rearranged location of functions
