+++ din_middleware.go
@@ -123,3 +123,17 @@ func (d *DinMiddleware) UnmarshalCaddyfile(dispenser *caddyfile.Dispenser) error
 	}
 	return nil
 }
+
+func urlToMetaUpstream(urlstr string) (*metaUpstream, error) {
+	url, err := url.Parse(urlstr)
+	if err != nil {
+		return nil, err
+	}
+	return &metaUpstream{
+		HttpUrl: urlstr,
+		path:    url.Path,
+		Headers: make(map[string]string),
+		// upstream: &reverseproxy.Upstream{Dial: fmt.Sprintf("%v://%v", url.Scheme, url.Host)},
+		upstream: &reverseproxy.Upstream{Dial: url.Host},
+	}, nil
+}
